project architecture update
