+++ Makefile
@@ -1,0 +1,5 @@
+up:
+	docker compose -f docker-compose.yml up -d
+
+down:
+	docker compose -f docker-compose.yml down && docker network prune --force
