+++ db/models.py
@@ -1,0 +1,19 @@
+import uuid
+
+from sqlalchemy import Boolean, Column, String
+from sqlalchemy.dialects.postgresql import UUID
+from sqlalchemy.orm import DeclarativeBase
+
+
+class Base(DeclarativeBase):
+    pass
+
+
+class User(Base):
+    __tablename__ = "users"
+
+    user_id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4())
+    name = Column(String, nullable=False)
+    surname = Column(String, nullable=False)
+    email = Column(String, nullable=False, unique=True)
+    is_active = Column(Boolean(), default=True)
