+++ docker-compose.yml
@@ -11,7 +11,19 @@ services:
       - 5432:5432
     networks:
       - custom
-
+  db_test:
+    container_name: db_test
+    image: postgres:latest # можно просто postgres - также будет браться последняя стабильная версия.
+    restart: always # Контейнер будет перезапускаться всегда
+    environment:
+      - POSTGRES_USER=postgres_test
+      - POSTGRES_PASSWORD=postgres_test
+      - POSTGRES_DB=postgres_test
+    ports:
+      - 5433:5432
+    networks:
+      - custom
+  
 
 networks:
   custom:
